Move get lisence and contributors into AboutUtil
